collapse classes and add back submissions
