+++ parsons/newmode/__init__.py
@@ -1,4 +1,3 @@
 from parsons.newmode.newmode import Newmode
-from parsons.newmode.newmode_v2 import NewmodeV2
 
-__all__ = ["Newmode", "NewmodeV2"]
+__all__ = ["Newmode"]
